Create To-Do List.txt
Time to get back to work...

First off, my last commit message was wrong; I'm still missing the treasure rooms, THEN it'll be gameplay-complete.

Second... wow, my last commit was 7 months ago. I was burned out on this then moved on to other projects, but I'm ready to continue this one, and I have a to-do list to (hopefully) prove it!

In case anyone actually reads these commit messages (and after this repo goes public), I've decided to split the remaining work into three releases:
- "Alpha", which will finish what I've already started and bring the game into a gameplay-complete state
- "Classic", which will recreate everything from the original Flash game, including the High Score feature that went unused in later versions
- "Reburninated", which will include all the new features I originally planned

After this, Reburninated may be ported to even more systems (no promises). At the very least, I want to run the SDL1 version of this game as a LibRetro core.