Text system overhaul
Reworked the text system (and polished other things) to utilize a new gameHiResScreen surface, which is the same physical size as the game screen but the highest possible resolution. This allows me to basically "cheat" the original game's resolution by having an additional screen that handles hi-res objects such as text (e.g. the game screen is always 250x180; but if that screen is scaled to 2x to fit a larger window, then the hi-res window is 500x360).

Additionally, text now automatically resizes when the window is resized. There are some very minor issues with this new system right now, but they will be fixed within the next few commits.