Assorted changes
- Added SDL_gfx for SDL1 (I forgot to do this last commit)
- Adjusted includes
- Adjusted Wii and Gamecube makefiles
- Set Wii/Gamecube screen size to 640x480
- Changed Gamecube rootDir

SDL_gfx's zoomSurface() is too slow to be used on every frame, even on a Switch. Thankfully, it won't be needed at all on SDL2 systems, but this is a problem for SDL1. I've come up with an alternate solution for SDL1 that uses zoomSurface() in a much more efficient way, but it'll take some time to implement.