Sprite system overhaul
Two major changes to the way graphics are handled.

First, The gameScreen, gameHiResScreen, and appScreen have been removed entirely. It was a good idea to separate different types of graphics into separate screens, but it ended up being much too intensive. In SDL1, three window-size (more or less) surfaces had to be blitted on every frame. SDL2 was even worse, with three window-size textures being CREATED on every frame. This is no longer needed, as everything is now handled using the windowScreen alone with changes being made to the outputRect each time a surface is blitted. This requires a decent amount of addition/multiplication, but that should be nothing compared to what was being done before.

Second, SDL2 now uses textures instead of surfaces, rendering each sprite directly to the renderer instead of using an intermediary screen. I thought I was being clever by utilizing one solution for both SDL1 and SDL2, but it was just too intensive in the end. Still, much of the sprite code is shared between both in this implementation.

These changes should make the game MUCH more efficient, and should probably fix the graphics issues the PSP version was having.