Sprite system overhaul part 1
Several big changes have been made to the sprite system, but none of them are actually being used in any meaningful capacity right now (that will come later, I just want to put out a functioning commit now).

SpriteObjects will use an array of SpriteSubObjects, which each contain a single sprite from a sprite sheet. A sprite sheet is processed to create standalone sprites of their exact sizes (as opposed to the current system, which requires each sprite to be the size of the largest sprite in the sheet). Not only will this reduce the RAM usage and per-frame processing of surfaces/textures by removing unnecessary pixel data, but it will also allow more information about each sprite to be stored, which will make the creation of cutscenes both simpler and more accurate. It also standardizes the creation of sprites among SDL2/SDL1 (normal)/SDL1 (zoomed), which is nice.