Statically linked fonts
Fonts are now statically linked to the executable instead of needing to be loaded from a separate directory. A similar change will later be made for all other assets.

Ultimately, this will simplify the process of distributing the compiled game since assets will be built into the executable. It also means per-system file IO will no longer need to be handled, most likely fixing the Gamecube port. It may also improve initial loading times for systems with slow file IO?

Unfortunately, my process for converting assets to C header files is functional but clunky, and the generated headers take up much more space than the original assets (this is unavoidable). I'll need to either include over half a gig of generated headers in the git repo, or figure out a way to include this process in the makefiles.