Partially implemented variable framerate support
Anything that relied on the original game's hardcoded 16 FPS (e.g. framerate variables, movement deltas) has been abstracted to use two variables: one that stores the true value, and another that rounds that value down to the nearest whole number. By utilizing this, animations and game speed are still consistent with the original game, while movement can smoothly match any framerate - even those that are not a multiple of 16.

This implementation is currently incomplete and broken on higher framerates, but the original 16 FPS still works fine, so I'm pushing it out now before this commit gets any bigger.