Updated GameManager for frame rate
Full disclosure: there's a bug that probably won't be fixed. Due to the way "enhanced" (variable) frame rate is adapted to original frame rate, some things that occur on a certain animation frame (e.g. ending the BURNINATE message animation and going back to the game) will occur up to one original frame, or up to 1/16 of a second, sooner that they would with a variable frame rate. Honestly, this bug is so minor and would be so much of a chore to rework that I'm not gonna bother. Also, it only affects framerates higher than the original 16 FPS.