FrameState overhaul
Frame state variables are now implemented as a new FrameState class, which stores the current frame state (both frame and subFrame), the next frame state, and indicators for whether the current subFrame is the start/end of the current frame. All existing behavior involving frames was then reworked to accommodate this new behavior (so it's a VERY big commit). This was necessary to fix many logic and sound issues, but it also fixed the problem I mentioned in the last commit that I said wouldn't be fixed.

Aside from cutscenes running too quickly at higher framerates and looking screwy as a result, the game appears to work just fine at any framerate; but I'll need to test it some more.