Made sprite color key optional
I was going to rework the sprite system to add the same improvements I made for SuDokuL, but this thing is a mess... that's what I get for trying to make the code TOO faithful to Flash ActionScript where sprites are center-based rather than top-left-based. I'll just cut my losses, this optimization should still be an improvement.